+++ cache/class.tx_rnbase_cache_Manager.php
@@ -22,7 +22,6 @@
  * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
  ***************************************************************/
 
-require_once(t3lib_extMgm::extPath('rn_base') . 'class.tx_rnbase.php');
 tx_rnbase::load('tx_rnbase_util_TYPO3');
 
 /**
@@ -30,7 +29,50 @@ tx_rnbase::load('tx_rnbase_util_TYPO3');
  */
 class tx_rnbase_cache_Manager {
 	private static $caches = array();
+	const CACHE_FRONTEND_VARIABLE = 'VariableFrontend';
+	const CACHE_FRONTEND_STRING = 'StringFrontend';
+	const CACHE_FRONTEND_PHP = 'PhpFrontend';
+	const CACHE_BACKEND_T3DATABASE = 'T3Database';
+	const CACHE_BACKEND_MEMCACHED = 'Memcached';
+	const CACHE_BACKEND_FILE = 'File';
+	const CACHE_BACKEND_REDIS = 'Redis';
+	const CACHE_BACKEND_APC = 'Apc';
+	const CACHE_BACKEND_PDO = 'Pdo';
+	const CACHE_BACKEND_TRANSIENTMEMORY = 'TransientMemory';
+	const CACHE_BACKEND_NULL = 'Null';
 
+	private static $aliases = array(
+		self::CACHE_FRONTEND_VARIABLE => array('TYPO3\CMS\Core\Cache\Frontend\VariableFrontend','t3lib_cache_frontend_VariableFrontend'),
+		self::CACHE_FRONTEND_STRING => array('TYPO3\CMS\Core\Cache\Frontend\StringFrontend','t3lib_cache_frontend_StringFrontend'),
+		self::CACHE_FRONTEND_PHP => array('TYPO3\CMS\Core\Cache\Frontend\PhpFrontend','t3lib_cache_frontend_PhpFrontend'),
+		self::CACHE_BACKEND_T3DATABASE => array('TYPO3\CMS\Core\Cache\Backend\Typo3DatabaseBackend','t3lib_cache_backend_DbBackend'),
+		self::CACHE_BACKEND_MEMCACHED => array('TYPO3\CMS\Core\Cache\Backend\MemcachedBackend','t3lib_cache_backend_MemcachedBackend'),
+		self::CACHE_BACKEND_FILE => array('TYPO3\CMS\Core\Cache\Backend\FileBackend','t3lib_cache_backend_FileBackend'),
+		self::CACHE_BACKEND_REDIS => array('TYPO3\CMS\Core\Cache\Backend\RedisBackend','t3lib_cache_backend_RedisBackend'),
+		self::CACHE_BACKEND_APC => array('TYPO3\CMS\Core\Cache\Backend\ApcBackend','t3lib_cache_backend_ApcBackend'),
+		self::CACHE_BACKEND_PDO => array('TYPO3\CMS\Core\Cache\Backend\PdoBackend','t3lib_cache_backend_PdoBackend'),
+		self::CACHE_BACKEND_TRANSIENTMEMORY => array('TYPO3\CMS\Core\Cache\Backend\TransientMemoryBackend','t3lib_cache_backend_TransientMemoryBackend'),
+		self::CACHE_BACKEND_NULL => array('TYPO3\CMS\Core\Cache\Backend\NullBackend','t3lib_cache_backend_NullBackend'),
+	);
+
+	/**
+	 * Register a TYPO3 cache
+	 * @param string $name
+	 * @param string $frontendKey see constants
+	 * @param string $backendKey see constants
+	 * @param array $options
+	 */
+	public static function registerCache($name, $frontendKey, $backendKey, $options = array()) {
+		$offset = tx_rnbase_util_TYPO3::isTYPO60OrHigher() ? 0 : 1;
+		$frontend = array_key_exists($frontendKey, self::$aliases) ? self::$aliases[$frontendKey][$offset] : $frontendKey;
+		$backend = array_key_exists($backendKey, self::$aliases) ? self::$aliases[$backendKey][$offset] : $backendKey;
+
+		$GLOBALS['TYPO3_CONF_VARS']['SYS']['caching']['cacheConfigurations'][$name] = array(
+				'frontend' => $frontend,
+				'backend' => $backend,
+				'options' => $options
+		);
+	}
 	/**
 	 * Liefert einen Cache
 	 *